logits to bf16

+++ src/transformers/models/llama/modeling_llama.py
@@ -1207,7 +1207,7 @@ class LlamaForCausalLM(LlamaPreTrainedModel):
                 )
             # Only compute necessary logits, and do not upcast them to float if we are not computing the loss
             # TODO: remove the float() operation in v4.46
-            logits = self.lm_head(hidden_states[:, -num_logits_to_keep:, :]).float()
+            logits = self.lm_head(hidden_states[:, -num_logits_to_keep:, :])
 
         loss = None
         if labels is not None:
